added admin page with full draft view (drafted and undrafted slots) and ability to create a draft (snake/straight, copy retainees or not)

+++ app/controllers/admin_controller.rb
@@ -3,29 +3,39 @@ class AdminController < ApplicationController
 
   def index
     # this is the control panel: buttons and status
+    @players_total_count = Player.count
+    @players_undrafted_count = Player.undrafted.count
+    @picks_total_count = Pick.count
+    @picks_undrafted_count = Pick.find_all_by_player_id("null").count
+    @picks = Pick.all
+  end
 
+  def new    
   end
 
   def create
-    users = User.all(:order => "draft_order asc")
-    if params[:draft_type] == "snake"
-      draft = set_snake(users, params[:rounds])
+    if !params[:rounds].nil? &&  (1..100).include?(params[:rounds].to_i)
+      users = User.all(:order => "draft_order asc")
+      if params[:draft_type] == "snake"
+        draft = set_snake(users, params[:rounds].to_i)
+      else
+        draft = set_straight(users,params[:rounds].to_i)
+      end
+      # destroy all current picks if drop_draft
+      if params[:drop_draft]
+        Pick.all.each do |x|
+          x.destroy
+        end
+      end
+      # create the new draft
+      make_it(draft)
+      if params[:copy_retainee]
+        copy_retainees_to_draft
+      end
     else
-      draft = set_straight(users,params[:rounds])
+      flash[:error] = "Rounds must be 1 and 100"
     end
-    # destroy all current picks
-    Pick.all.each do |x|
-      x.destroy
-    end
-    # create the new draft
-    make_it(draft)
-    # find all draft picks to show admin
-  end
-
-  def yahoo
-  end
-
-  def previous
+    redirect_to :action => "index"
   end
 
   private
@@ -58,7 +68,7 @@ class AdminController < ApplicationController
     return draft
   end
 
-  def straight(users, rounds)
+  def set_straight(users, rounds)
     # users is a array of user objects in draft_pick asc order
     league_size = users.size
     z = 0
@@ -80,8 +90,14 @@ class AdminController < ApplicationController
     end
   end
   
-  def show_git_version
-    return system("git rev-parse HEAD")
+  def copy_retainees_to_draft
+    retainees = Retainee.all
+    retainees.each do |r|
+      draft_choice = Pick.remaining.first
+      draft_choice.player_id = r.player_id
+      draft_choice.user_id = r.user_id
+      draft_choice.save
+    end
   end
-
+  
 end
